feat: KeyboardShortcuts added for toolbar, style: Move ViewModifiers into a new file
